+++ music_assistant/controllers/music/albums.py
@@ -7,13 +7,13 @@ from typing import List
 from music_assistant.constants import EventType
 from music_assistant.helpers.cache import cached
 from music_assistant.helpers.compare import compare_album, compare_strings
-from music_assistant.helpers.util import create_sort_name, merge_dict, merge_list
 from music_assistant.helpers.json import json_serializer
+from music_assistant.helpers.util import create_sort_name, merge_dict, merge_list
+from music_assistant.models.media_controller import MediaControllerBase
 from music_assistant.models.media_items import (
     Album,
     AlbumType,
     ItemMapping,
-    MediaControllerBase,
     MediaType,
     Track,
 )
