+++ music_assistant/controllers/music/artists.py
@@ -11,14 +11,14 @@ from music_assistant.helpers.compare import (
     compare_strings,
     compare_track,
 )
-from music_assistant.helpers.util import create_sort_name, merge_dict, merge_list
 from music_assistant.helpers.json import json_serializer
+from music_assistant.helpers.util import create_sort_name, merge_dict, merge_list
+from music_assistant.models.media_controller import MediaControllerBase
 from music_assistant.models.media_items import (
     Album,
     AlbumType,
     Artist,
     ItemMapping,
-    MediaControllerBase,
     MediaType,
     Track,
 )
