+++ music_assistant/controllers/music/tracks.py
@@ -12,9 +12,9 @@ from music_assistant.helpers.compare import (
 )
 from music_assistant.helpers.util import create_sort_name, merge_dict, merge_list
 from music_assistant.helpers.json import json_serializer
+from music_assistant.models.media_controller import MediaControllerBase
 from music_assistant.models.media_items import (
     ItemMapping,
-    MediaControllerBase,
     MediaType,
     Track,
 )
